IpickSOC-3692 add code scanning alerts backfill
