+++ src/github/client/github-client.types.ts
@@ -52,6 +52,7 @@ export type GraphQlQueryResponse<ResponseData> = {
 };
 
 export type PaginatedAxiosResponse<T> = { hasNextPage: boolean; } & AxiosResponse<T>;
+
 export interface ReposGetContentsResponse {
 	content?: string;
 	download_url: string | null;
@@ -66,7 +67,9 @@ export interface ReposGetContentsResponse {
 	url: string;
 }
 
-export type ActionsListRepoWorkflowRunsResponseEnhanced = Octokit.ActionsListRepoWorkflowRunsResponse & {name: string};
+export type ActionsListRepoWorkflowRunsResponseEnhanced =
+	Octokit.ActionsListRepoWorkflowRunsResponse
+	& { name: string };
 
 
 export type CreateReferenceBody = {
@@ -110,16 +113,60 @@ export type DependabotAlertResponseItem = {
 }
 
 export type SecretScanningAlertResponseItem = {
-    number: number,
-    created_at: string,
-    updated_at?: string,
-    url: string,
-    html_url: string,
-    locations_url: string,
-    state: "open" | "resolved",
-    resolution?: string,
-    resolved_at?: string,
-    resolution_comment?: string,
-    secret_type: string,
-    secret_type_display_name: string
-  }
+	number: number,
+	created_at: string,
+	updated_at?: string,
+	url: string,
+	html_url: string,
+	locations_url: string,
+	state: "open" | "resolved",
+	resolution?: string,
+	resolved_at?: string,
+	resolution_comment?: string,
+	secret_type: string,
+	secret_type_display_name: string
+}
+
+export type CodeScanningAlertResponseItem = {
+	number: number;
+	created_at: string;
+	updated_at: string;
+	url: string;
+	html_url: string;
+	instances_url: string;
+	state: "open" | "dismissed" | "fixed";
+	fixed_at: string;
+	dismissed_at: string;
+	dismissed_reason: null | "false positive" | "won't fix" | "used in tests";
+	dismissed_comment: string;
+	rule: CodeScanningAlertResponseItemRule;
+	tool: CodeScanningAlertResponseItemTool;
+	most_recent_instance: CodeScanningAlertResponseItemMostRecentInstance;
+};
+
+type CodeScanningAlertResponseItemRule = {
+	name: string;
+	description: string;
+	full_description: string;
+	id: string | null;
+	tags: string[] | null;
+	severity: "none" | "note" | "warning" | "error" | null;
+	security_severity_level: "low" | "medium" | "high" | "critical" | null;
+	help: string | null;
+	help_uri: string | null;
+};
+
+type CodeScanningAlertResponseItemTool = {
+	name: string;
+	version: string | null;
+	guid: string | null;
+}
+
+type CodeScanningAlertResponseItemMostRecentInstance = {
+	ref: string;
+	environment: string;
+	category: string;
+	state: string;
+	commit_sha: string;
+	html_url: string;
+}
